agents/judge/pyproject.toml
[project]
name = "judge-agent"
version = "0.1.0"
description = "Judge microservice agent"
requires-python = ">=3.13"
dependencies = [
    "shared-utils @ file:///${PROJECT_ROOT}/shared",
    "google-adk",
    "a2a-sdk",
    "fastapi",
    "uvicorn",
    "python-json-logger",
    "google-cloud-logging",
    "opentelemetry-exporter-gcp-trace",
]

[tool.hatch.build.targets.wheel]
packages = ["agent"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
